solving for task 7 to 12

+++ 0x00-MySQL_Advanced/10-div.sql
@@ -1,7 +1,7 @@
 -- a SQL script that creates a function SafeDiv that divides (and returns) the first by the second number or returns 0 if the second number is equal to 0.
 
 CREATE FUNCTION SafeDiv(a INT, b INT)
-RETURNS INT
+RETURNS FLOAT
 BEGIN
     IF b = 0 THEN
         RETURN 0;
